Nyomdakönyvhöz 3 kép feltöltésének lehetősége.
Mostantól legfeljebb 3 kép tölthető fel egy nyomdakönyvi munkához. Forgathatók, törölhetők, e-mailben küldhetők.
Ajax-os feltöltéssel lett megoldva, hogy ne legyen oldal újratöltés.

+++ protected/extensions/EFineUploader/qqFileUploader.php
@@ -263,7 +263,7 @@ class qqFileUploader {
      * @param string $str
      */
     protected function toBytes($str){
-        $val = trim($str);
+        $val = (int)trim($str);
         $last = strtolower($str[strlen($str)-1]);
         switch($last) {
             case 'g': $val *= 1024;
